openscoring-service/src/test/java/org/openscoring/service/ModelRefTest.java
/*
 * Copyright (c) 2019 Villu Ruusmann
 *
 * This file is part of Openscoring
 *
 * Openscoring is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * Openscoring is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with Openscoring.  If not, see <http://www.gnu.org/licenses/>.
 */
package org.openscoring.service;

import org.junit.Test;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

public class ModelRefTest {

	@Test
	public void validateId(){
		assertTrue(ModelRef.validateId("com.mycompany.Test"));

		assertTrue(ModelRef.validateId("Test_v1"));
		assertTrue(ModelRef.validateId("Test_v1.0"));

		assertFalse(ModelRef.validateId("."));
		assertFalse(ModelRef.validateId("-Test"));
	}
}